Final V2

+++ view/zona.usuarios.php
@@ -1,9 +1,10 @@
 <?php
 require_once "../services/connection.php";
 session_start();
-if ($_SESSION['email']=="") {
+if (!isset($_SESSION['email']) || $_SESSION['email']=="" && !isset($_SESSION['tipo_u']) || $_SESSION['tipo_u']=="") {
     header("location:login.php");
 }else {
+    $id_usuario=$_SESSION['id_u'];
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -83,44 +84,47 @@ if ($_SESSION['email']=="") {
             $data[]="disponibilidad_u = '{$disponibilidad_u}'";
         }
         $anadir= implode(' AND ',$data);
-        $usuarios=$pdo->prepare("SELECT * FROM tbl_usuario WHERE {$anadir}");
+        if (!empty($data)){
+            $usuarios=$pdo->prepare("SELECT * FROM tbl_usuario WHERE {$anadir}");
+        }else{
+            $usuarios=$pdo->prepare("SELECT * FROM tbl_usuario");
+        }
         try{
             $pdo->beginTransaction();
             $usuarios->execute();
-            if (empty($filtrar)) {
-                echo "<div>";
-                echo "<h1>No se han encontrado elementos....</h1>";
-                echo "</div>";
-            }else {
-                echo  "<div>";
-                echo  "<table>";
+            echo  "<div>";
+            echo  "<table>";
+            echo  "<tr>";
+            echo  "<th class='blue'>Nombre</th>";
+            echo  "<th class='blue'>Apellido</th>";
+            echo  "<th class='blue'>Correo</th>";
+            echo  "<th class='blue'>Tipo</th>";
+            echo  "<th class='blue'>Disponibilidad</th>";
+            echo  "</tr>";
+            foreach ($usuarios as $usuario) {
+                //Ponemos primero la localización
                 echo  "<tr>";
-                echo  "<th class='blue'>Nombre</th>";
-                echo  "<th class='blue'>Apellido</th>";
-                echo  "<th class='blue'>Correo</th>";
-                echo  "<th class='blue'>Tipo</th>";
-                echo  "<th class='blue'>Disponibilidad</th>";
-                echo  "</tr>";
-                foreach ($usuarios as $usuario) {
-                    //Ponemos primero la localización
-                    echo  "<tr>";
-                        echo "<td class='gris'>{$usuario['nombre_u']}</td>";
-                        echo "<td class='gris'>{$usuario['apellido_u']}</td>";
-                        echo "<td class='gris'>{$usuario['correo_u']}</td>";
-                        echo "<td class='gris'>{$usuario['tipo_u']}</td>";
-                        if ($usuario['disponibilidad_u']=="si") {
-                            echo "<td class='gris'><i class='fas fa-check green'></i></td>";
-                        }else{
-                            echo "<td class='gris'><i class='fas fa-times red'></i></td>";
-                        }
+                    echo "<td class='gris'>{$usuario['nombre_u']}</td>";
+                    echo "<td class='gris'>{$usuario['apellido_u']}</td>";
+                    echo "<td class='gris'>{$usuario['correo_u']}</td>";
+                    echo "<td class='gris'>{$usuario['tipo_u']}</td>";
+                    if ($usuario['disponibilidad_u']=="si") {
+                        echo "<td class='gris'><i class='fas fa-check green'></i></td>";
+                    }else{
+                        echo "<td class='gris'><i class='fas fa-times red'></i></td>";
+                    }
+                    if ($usuario['id_u']==$id_usuario) {
+                        echo "<td><button type='submit' class='red'>Cambia de cuenta para modificar esta</button></td>";
+                        echo "<td><button type='submit' class='red'>Cambia de cuenta para eliminar esta</button></td>";
+                    }else{
                         echo "<td><button type='submit'><a type='button' href='mod.usuarios.php?id_u={$usuario['id_u']}'>Modificar usuario</a></button></td>";
                         echo "<td><button type='submit'><a type='button' href='../proceses/eliminarusuario.php?id_u={$usuario['id_u']}'>Eliminar usuario</a></button></td>";
-                    echo "</tr>";
-                }
-                echo "</table>";
-                echo "</div>";
-                $pdo->commit();
+                    }
+                echo "</tr>";
             }
+            echo "</table>";
+            echo "</div>";
+            $pdo->commit();
         } catch (Exception $e) {
             $pdo->rollBack();
             echo "Fallo: " . $e->getMessage();
@@ -152,8 +156,13 @@ if ($_SESSION['email']=="") {
                     }else{
                         echo "<td class='gris'><i class='fas fa-times red'></i></td>";
                     }
-                    echo "<td><button type='submit'><a type='button' href='mod.usuarios.php?id_u={$usuario['id_u']}'>Modificar usuario</a></button></td>";
-                    echo "<td><button type='submit'><a type='button' href='../proceses/eliminarusuario.php?id_u={$usuario['id_u']}'>Eliminar usuario</a></button></td>";
+                    if ($usuario['id_u']==$id_usuario) {
+                        echo "<td><button type='submit' class='red'>Cambia de cuenta para modificar esta</button></td>";
+                        echo "<td><button type='submit' class='red'>Cambia de cuenta para eliminar esta</button></td>";
+                    }else{
+                        echo "<td><button type='submit'><a type='button' href='mod.usuarios.php?id_u={$usuario['id_u']}'>Modificar usuario</a></button></td>";
+                        echo "<td><button type='submit'><a type='button' href='../proceses/eliminarusuario.php?id_u={$usuario['id_u']}'>Eliminar usuario</a></button></td>";
+                    }
                 echo "</tr>";
             }
             echo "</table>";
